remove .idea, a.out and  cmake-buil-debug from git

+++ main.c
@@ -1,5 +1,4 @@
 #include <stdio.h>
-#include <stdbool.h>
 #include <string.h>
 #include <stdlib.h>
 #include <errno.h>
@@ -11,9 +10,12 @@
 #include <signal.h>
 #include <sys/time.h>
 
+// Вариант 10  Дерево процессов: 1->2   2->3   3->(4,5,6)    6->7  4->8
+// Вариант 11  Последовательность обмена сигналами: 1->(8,7) SIGUSR1   8->(6,5) SIGUSR1  5->(4,3,2) SIGUSR2   2->1 SIGUSR2
+
 char *module;
 int X = 0, Y = 0;
-pid_t pid0 = 0, pid1 = 0, pid2 = 0, pid3 = 0, pid4 = 0, pid5 = 0, pid6 = 0, pid7 = 0, pid8 = 0, group_pid = 0;
+pid_t pid0 = 0, pid1 = 0, pid2 = 0, pid3 = 0, pid4 = 0, pid5 = 0, pid6 = 0, pid7 = 0, pid8 = 0;
 
 void printErr(const char *module, const char *errmsg, const char *filename) {
     fprintf(stderr, "%d %s: %s %s\n", getpid(), module, errmsg, filename ? filename : "");
@@ -25,35 +27,17 @@ long getTime() {
     return (tv.tv_usec) % 1000;
 }
 
-void debug() {
-    printf("PID^ %d", getpid());
-    printf("\n%d %d", 1, pid1);
-    printf("\n%d %d", 2, pid2);
-    printf("\n%d %d", 3, pid3);
-    printf("\n%d %d", 4, pid4);
-    printf("\n%d %d", 5, pid5);
-    printf("\n%d %d", 6, pid6);
-    printf("\n%d %d", 7, pid7);
-    printf("\n%d %d", 8, pid8);
-}
-
 void handler1() {
-    static int recieved = 0;
-    recieved++;
+    static int received = 0;
+    received++;
 
-    printf("1 %d %d получил USR1 %ld\n", getpid(), getppid(), getTime());
+    printf("1 %d %d получил USR2 %ld\n", getpid(), getppid(), getTime());
 
-    if (recieved == 101) {
-        //sleep(5);
-
-        char buffer[256];
-        sprintf(buffer, "pstree -p %d", getppid());
-        system(buffer);
-
-
+    if (received == 101) {
+        sleep(1);
         if (kill(-pid7, SIGTERM) == -1)
             printErr(module, strerror(errno), NULL);
-        int children = 1;
+        int children = 2;
         while (children--)
             wait(NULL);
         printf("%d %d завершил работу после %d-го сигнала SIGUSR1 и %d-го сигнала SIGUSR2\n", getpid(), getppid(), X,
@@ -65,9 +49,7 @@ void handler1() {
         } else {
             printf("1 %d %d послал USR1 %ld\n", getpid(), getppid(), getTime());
         }
-        //printf("1 %d %d послал USR1 %ld\n", getpid(), getppid(), getTime());
         X++;
-
     }
 }
 
@@ -82,20 +64,10 @@ void handler2() {
 
 void handler3() {
     printf("3 %d %d получил USR2 %ld\n", getpid(), getppid(), getTime());
-//    if (kill(pid6, SIGUSR1) == -1)
-//        printErr(module, strerror(errno), NULL);
-//    else
-//        printf("3 %d %d послал USR1 %ld\n", getpid(), getppid(), getTime());
-//    X++;
 }
 
 void handler4() {
     printf("4 %d %d получил USR2 %ld\n", getpid(), getppid(), getTime());
-//    if (kill(pid5, SIGUSR1) == -1)
-//        printErr(module, strerror(errno), NULL);
-//    else
-//        printf("4 %d %d послал USR1 %ld\n", getpid(), getppid(), getTime());
-//    X++;
 }
 
 void handler5() {
@@ -103,26 +75,16 @@ void handler5() {
     if (kill(-pid2, SIGUSR2) == -1)
         printErr(module, strerror(errno), NULL);
     else
-        printf("5 %d %d послал USR3 %ld\n", getpid(), getppid(), getTime());
+        printf("5 %d %d послал USR2 %ld\n", getpid(), getppid(), getTime());
     Y++;
 }
 
 void handler6() {
     printf("6 %d %d получил USR1 %ld\n", getpid(), getppid(), getTime());
-//    if (kill(pid7, SIGUSR1) == -1)
-//        printErr(module, strerror(errno), NULL);
-//    else
-//        printf("6 %d %d послал USR1 %ld\n", getpid(), getppid(), getTime());
-//    X++;
 }
 
 void handler7() {
     printf("7 %d %d получил USR1 %ld\n", getpid(), getppid(), getTime());
-//    if (kill(pid8, SIGUSR1) == -1)
-//        printErr(module, strerror(errno), NULL);
-//    else
-//        printf("7 %d %d послал USR1 %ld\n", getpid(), getppid(), getTime());
-//    X++;
 }
 
 void handler8() {
@@ -135,31 +97,16 @@ void handler8() {
 }
 
 void handlerTerm2() {
-    if (kill(pid1, SIGTERM) == -1)
-        printErr(module, strerror(errno), NULL);
-    int children = 1;
-    while (children--)
-        wait(NULL);
     printf("%d %d завершил работу после %d-го сигнала SIGUSR1 и %d-го сигнала SIGUSR2\n", getpid(), getppid(), X, Y);
     exit(0);
 }
 
 void handlerTerm3() {
-//    if (kill(pid6, SIGTERM) == -1)
-//        printErr(module, strerror(errno), NULL);
-//    int children = 1;
-//    while(children--)
-//        wait(NULL);
     printf("%d %d завершил работу после %d-го сигнала SIGUSR1 и %d-го сигнала SIGUSR2\n", getpid(), getppid(), X, Y);
     exit(0);
 }
 
 void handlerTerm4() {
-//    if (kill(pid5, SIGTERM) == -1)
-//        printErr(module, strerror(errno), NULL);
-//    int children = 1;
-//    while(children--)
-//        wait(NULL);
     printf("%d %d завершил работу после %d-го сигнала SIGUSR1 и %d-го сигнала SIGUSR2\n", getpid(), getppid(), X, Y);
     exit(0);
 }
@@ -175,27 +122,21 @@ void handlerTerm5() {
 }
 
 void handlerTerm6() {
-//    if (kill(pid7, SIGTERM) == -1)
-//        printErr(module, strerror(errno), NULL);
-//    int children = 1;
-//    while (children--)
-//        wait(NULL);
     printf("%d %d завершил работу после %d-го сигнала SIGUSR1 и %d-го сигнала SIGUSR2\n", getpid(), getppid(), X, Y);
     exit(0);
 }
 
 void handlerTerm7() {
-//    if (kill(pid8, SIGTERM) == -1)
-//        printErr(module, strerror(errno), NULL);
-//    int children = 1;
-//    while (children--)
-//        wait(NULL);
     printf("%d %d завершил работу после %d-го сигнала SIGUSR1 и %d-го сигнала SIGUSR2\n", getpid(), getppid(), X, Y);
     exit(0);
 }
 
 void handlerTerm8() {
-    //  потому что последний
+    if (kill(-pid5, SIGTERM) == -1)
+        printErr(module, strerror(errno), NULL);
+    int children = 2;
+    while (children--)
+        wait(NULL);
     printf("%d %d завершил работу после %d-го сигнала SIGUSR1 и %d-го сигнала SIGUSR2\n", getpid(), getppid(), X, Y);
     exit(0);
 }
@@ -241,63 +182,67 @@ void makeTree() {
         pid2 = fork();
         if (!pid2)
             pid2 = getpid();
-        else if (pid2 < 0)
+        else if (pid2 < 0) {
             printErr(module, strerror(errno), NULL);
-    } else if (pid1 < 0)
+        }
+    } else if (pid1 < 0) {
         printErr(module, strerror(errno), NULL);
-
+    }
 
     if (getpid() == pid2) {
         pid3 = fork();
-        if (!pid3)
+        if (!pid3) {
             pid3 = getpid();
-        else if (pid3 < 0)
+        } else if (pid3 < 0) {
             printErr(module, strerror(errno), NULL);
+        }
     }
-
 
     if (getpid() == pid3) {
         pid4 = fork();
-        if (!pid4)
+        if (!pid4) {
             pid4 = getpid();
-        else if (pid4 < 0)
+        } else if (pid4 < 0) {
             printErr(module, strerror(errno), NULL);
+        }
     }
 
     if (getpid() == pid3) {
         pid5 = fork();
-        if (!pid5)
+        if (!pid5) {
             pid5 = getpid();
-        else if (pid5 < 0)
+        } else if (pid5 < 0) {
             printErr(module, strerror(errno), NULL);
+        }
     }
 
     if (getpid() == pid3) {
         pid6 = fork();
-        if (!pid6)
+        if (!pid6) {
             pid6 = getpid();
-        else if (pid6 < 0)
+        } else if (pid6 < 0) {
             printErr(module, strerror(errno), NULL);
+        }
     }
 
     if (getpid() == pid4) {
-        pid7 = fork();
-        if (!pid7) {
-            pid7 = getpid();
-        }
-        else if (pid7 < 0)
+        pid8 = fork();
+        if (!pid8) {
+            pid8 = getpid();
+        } else if (pid8 < 0) {
             printErr(module, strerror(errno), NULL);
+        }
     }
 
     if (getpid() == pid6) {
-        pid8 = fork();
-        if (!pid8)
-            pid8 = getpid();
-        else if (pid8 < 0)
+        pid7 = fork();
+        if (!pid7) {
+            pid7 = getpid();
+        } else if (pid7 < 0) {
             printErr(module, strerror(errno), NULL);
+        }
     }
 }
-
 
 void exchangeSignals(const char *dirname) {
     struct sigaction sg;
@@ -308,13 +253,11 @@ void exchangeSignals(const char *dirname) {
     sigemptyset(&sgTerm.sa_mask);
     sgTerm.sa_flags = SA_SIGINFO;
 
-
     if (getpid() == pid1) {
         createFile('1');
 
         const int required = 8;
         int curr = 0;
-        //todo ?????????????????????
         while (curr != required) {
             curr = 0;
             DIR *currdir;
@@ -340,6 +283,14 @@ void exchangeSignals(const char *dirname) {
         pid7 = getPidFromFile('7');
         pid8 = getPidFromFile('8');
 
+        sg.sa_sigaction = handler1;
+        if (sigaction(SIGUSR2, &sg, NULL) == -1) {
+            printErr(module, strerror(errno), NULL);
+            exit(1);
+        }
+        char buffer[256];
+        sprintf(buffer, "pstree -p %d", getppid());
+        system(buffer);
         if (kill(-pid7, SIGUSR1) == -1)
             printErr(module, strerror(errno), NULL);
         else
@@ -350,9 +301,8 @@ void exchangeSignals(const char *dirname) {
 
 //    debug();
     if (getpid() == pid2) {
-        printf("TUT");
         sg.sa_sigaction = handler2;
-        if (sigaction(SIGUSR1, &sg, NULL) == -1) {
+        if (sigaction(SIGUSR2, &sg, NULL) == -1) {
             printErr(module, strerror(errno), NULL);
             exit(1);
         }
@@ -402,7 +352,7 @@ void exchangeSignals(const char *dirname) {
             printErr(module, strerror(errno), NULL);
             exit(1);
         }
-//        pid5 = getPidFromFile('5');
+
         createFile('4');
     }
 
@@ -426,7 +376,7 @@ void exchangeSignals(const char *dirname) {
         createFile('5');
     }
 
-    //todo
+
     if (getpid() == pid6) {
         sg.sa_sigaction = handler6;
         if (sigaction(SIGUSR1, &sg, NULL) == -1) {
@@ -438,11 +388,13 @@ void exchangeSignals(const char *dirname) {
             printErr(module, strerror(errno), NULL);
             exit(1);
         }
-        pid6 = getPidFromFile('6');
+
+        pid5 = getPidFromFile('5');
         if (setpgid(pid6, pid5) == -1) {
             printErr(module, strerror(errno), NULL);
             exit(1);
         }
+
         createFile('6');
     }
 
@@ -514,33 +466,19 @@ void exchangeSignals(const char *dirname) {
             printErr(module, strerror(errno), dirname);
             exit(1);
         }
-    } else/*{
-        pid_t myPid = getpid();
-        int me = 0;
-        if (pid1 == myPid) me = 1;
-        if (pid2 == myPid) me = 2;
-        if (pid3 == myPid) me = 3;
-        if (pid4 == myPid) me = 4;
-        if (pid5 == myPid) me = 5;
-        if (pid6 == myPid) me = 6;
-        if (pid7 == myPid) me = 7;
-        if (pid8 == myPid) me = 8;
-        printf("я %d %d %d\n", me, getpid(), getpgid(0));
+    } else {
+        while (1) pause();
     }
-        */while (1) pause();
 }
 
 
 int main(int argc, char *argv[]) {
-    printf("TUT");
     module = basename(argv[0]);
     char *dirname = "/tmp/lab4";
-    printf("TUT");
     if (mkdir(dirname, 0777) == -1) {
         printErr(module, strerror(errno), dirname);
         exit(1);
     }
-    printf("TUT");
     makeTree();
     exchangeSignals(dirname);
     exit(0);
